# Conflicts:
#	readme.md

+++ Samples/luainvaders/src/Defender.h
@@ -69,8 +69,8 @@ private:
 	bool bullet_in_flight;
 	Rml::Core::Vector2f bullet_position;
 
-	float defender_frame_start;
-	float respawn_start;
+	double defender_frame_start;
+	double respawn_start;
 
 	bool render;
